+++ intern1/include/Animation.h
@@ -8,8 +8,8 @@ class AnimationTreeTraversal;
 
 struct AnimationData
 {
-        std::vector<std::pair<float,glm::quat>>     _localRotation;
-        std::vector<std::pair<float,glm::mat4>>     _localTrans;
+        std::vector<std::pair<uint32,glm::quat>>     _localRotation;
+        std::vector<std::pair<uint32,glm::mat4>>     _localTrans;
 
         int32                                       _boneIndex;
         std::vector<AnimationData>                  _childrens;
