+++ intern1/include/Simulator.h
@@ -16,8 +16,8 @@ class Simulator : Noncopyable
         std::vector<glm::mat4> _transForm;
         std::vector<uint32>    _compresskeyFrame[4];//애니메이션 개수 만큼 배열 되어야 할듯
 
-        void updateTransForm(const AnimationData& node, glm::mat4 wolrdTrans, float keyTime);
-        void getFrameIterator(float* keyArray, float findKeyTime, const std::vector<std::pair<float,glm::quat>>& animationFrame);
+        void updateTransForm(const AnimationData& node, glm::mat4 wolrdTrans, uint32 keyTime);
+        void getFrameIterator(uint32* keyArray, uint32 findKeyTime, const std::vector<std::pair<uint32,glm::quat>>& animationFrame);
     public:
         Simulator()
         {
